+++ Changelog.md
@@ -6,6 +6,8 @@ _2024.01.10_
 ### New features
 
 ### Changes
+* Add a new parameter to select the number of roots to initialise a TPG
+
 * Add two parameters to the reset method of the learning environment. These parameters are used for environments that use specific initialization.
   * Parameter `iterationNumber`: an integer indicating the current iteration number when the `nbIterationsPerPolicyEvaluation` parameter is greater than 1, default value = 0.
   * Parameter `generationNumber`: an integer indicating the current generation number, default value = 0.